fix(flow-next): skip cleanup in CREATE mode

+++ plugins/flow-next/scripts/ralph_smoke_rp.sh
@@ -107,7 +107,7 @@ find_jsonl() {
 }
 
 cleanup() {
-  if [[ "${KEEP_TEST_DIR:-0}" != "1" ]]; then
+  if [[ "${KEEP_TEST_DIR:-0}" != "1" && "${CREATE:-0}" != "1" ]]; then
     rm -rf "$TEST_DIR"
   fi
 }
